requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anykeystore==0.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
Babel==2.9.0
backcall==0.2.0
bce-python-sdk==0.8.53
beautifulsoup4==4.11.1
bert-base==0.0.9
bert-serving==0.0.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bert4keras==0.10.8
bleach==5.0.0
boto3==1.21.43
botocore==1.24.43
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
cfgv==3.2.0
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
charset-normalizer==2.0.12
cleanlab==1.0
click==7.1.2
colorama==0.4.4
colorlog==6.6.0
cryptacular==1.6.2
cryptography @ file:///tmp/build/80754af9/cryptography_1605544449973/work
cycler==0.10.0
debugpy==1.6.0
decorator==4.4.2
decord==0.5.2
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.1
dpcpp-cpp-rt==2022.1.0
easydict==1.9
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints==0.3
fastjsonschema==2.15.3
ffmpeg==1.4
filelock==3.0.12
fire==0.3.1
flake8==3.8.4
Flask-Babel==2.0.0
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gast==0.2.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==1.1.2
grpcio==1.33.2
gunicorn==20.0.4
h5py==3.7.0
huggingface-hub==0.4.0
hupper==1.10.3
identify==1.5.10
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
idna-ssl==1.1.0
imageio==2.16.0
imageio-ffmpeg==0.4.7
imbalanced-learn==0.7.0
imgaug==0.4.0
importlib-metadata==4.11.4
importlib-resources==5.7.1
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
ipykernel==6.13.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonlines==3.0.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.38.0
lmdb==1.0.0
ltp==4.0.9
lxml==4.6.2
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl==2022.1.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
multidict==5.2.0
multiprocess==0.70.12.2
nbclient==0.5.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.1
networkx==2.5
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
nodeenv==1.5.0
notebook==6.4.11
numba==0.55.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
objgraph @ file:///package/objgraph-3.4.1.tar.gz
olefile==0.46
opencv-python~=4.2.0.32
opt-einsum==3.3.0
packaging==21.3
paddlehub==2.0.0b0
paddleocr==1.1.1
paddlepaddle-gpu==2.0.2
pandas~=1.2.4
pandocfilters==1.5.0
parso==0.8.1
PasteDeploy==2.1.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow~=9.2.0
plaster==1.0
plaster-pastedeploy==0.7
pre-commit==2.9.2
prettytable @ file:///package/prettytable-0.7.tar.gz
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.14.0
psutil==5.9.1
ptyprocess==0.7.0
py-cpuinfo==5.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.0
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.9.9
pyflakes==2.2.0
Pygments==2.8.0
pygtrie==2.3.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==2.1.0
rarfile @ file:///package/rarfile-3.0.tar.gz
regex~=2022.7.9
repoze.sendmail==4.4.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.1
resampy==0.2.2
responses==0.17.0
rsa==4.6
s3transfer==0.5.2
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn~=0.24.1
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
Send2Trash==1.8.0
sentencepiece==0.1.94
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1605205313296/work
smart-open==5.0.0
smmap==3.0.4
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.31
tbb==2021.6.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.9.2
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.5.4
tinycss2==1.1.1
tokenizers==0.8.1rc2
toml==0.10.2
torch==1.6.0
torchtext==0.5.0
torchvision==0.7.0
tornado==6.0.4
tqdm~=4.64.0
traitlets==4.3.3
transaction==3.0.1
transformers==3.1.0
translationstring==1.4
typing-extensions==4.0.1
ujson==4.3.0
unzip==1.0.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
velruse==1.1.1
venusian==3.0.0
virtualenv==20.2.1
visualdl==2.0.5
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==3.0.0
wtforms-recaptcha==0.3.2
xlrd==1.2.0
xxhash==3.0.0
yarl==1.7.2
zipp==3.4.0
zope.deprecation==4.4.0
zope.interface==5.4.0

pymongo~=3.11.3
numpy~=1.21.6
Flask~=1.1.2
requests~=2.24.0
ijson~=3.1.4
six~=1.15.0
odps~=3.5.1
gradio~=3.5
waitress~=1.4.4